fix: iOS/tvOS native module linking and bridge configuration
- Fix react-native.config.json: use 'dependency' format instead of
  'platforms' array (was being treated as a Platform Provider, not a
  dependency, causing iOS autolinking to skip the module entirely)
- Fix podspec path reference from ./ios/ to ./ (podspec lives at root)
- Replace deprecated 'React-Core' dependency with install_modules_dependencies()
  for proper RN 0.81+ New Architecture compatibility
- Fix critical ObjC/Swift name mismatch: bridge declared MpvPlayerViewManager
  but Swift class was @objc(MpvViewManager) — renamed to MpvPlayerViewManager
  so RN correctly derives view name 'MpvPlayerView'
- Add missing Bridging Header (react-native-mpv-Bridging-Header.h)
- Fix ObjC bridge trackId type from int to NSInteger (matches Swift Int)
- Raise deployment target from 15.1 to 16.0 for MPVKit compatibility
- Remove direct 'import React' from Swift files (use bridging header instead)

+++ modules/react-native-mpv/ios/MpvViewManagerBridge.m
@@ -1,4 +1,5 @@
 #import <React/RCTViewManager.h>
+#import <React/RCTBridgeModule.h>
 
 @interface RCT_EXTERN_MODULE(MpvPlayerViewManager, RCTViewManager)
 
@@ -16,8 +17,8 @@ RCT_EXPORT_VIEW_PROPERTY(onMpvTracksChanged, RCTDirectEventBlock)
 
 RCT_EXTERN_METHOD(seekTo:(nonnull NSNumber *)node seconds:(double)seconds)
 RCT_EXTERN_METHOD(seekRelative:(nonnull NSNumber *)node seconds:(double)seconds)
-RCT_EXTERN_METHOD(setAudioTrack:(nonnull NSNumber *)node trackId:(int)trackId)
-RCT_EXTERN_METHOD(setSubtitleTrack:(nonnull NSNumber *)node trackId:(int)trackId)
+RCT_EXTERN_METHOD(setAudioTrack:(nonnull NSNumber *)node trackId:(NSInteger)trackId)
+RCT_EXTERN_METHOD(setSubtitleTrack:(nonnull NSNumber *)node trackId:(NSInteger)trackId)
 RCT_EXTERN_METHOD(stop:(nonnull NSNumber *)node)
 
 @end
